.ipynb_checkpoints/requirements-checkpoint.txt
streamlit==1.28.1
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1